requirements.txt
python-telegram-bot~=21.10
sqlalchemy~=2.0.37
apscheduler~=3.11.0
python-dotenv~=1.0.1
python-dateutil~=2.9.0.post0